task-manager-flask/requirements.txt
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-datastore==2.6.1
googleapis-common-protos==1.56.2
greenlet==1.1.2
grpcio==1.46.3
grpcio-status==1.46.3
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
proto-plus==1.20.5
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
